cleaned up list
